requirements.txt
Django==5.1.2
djangorestframework==3.15.2
django-cors-headers==4.3.1
python-dotenv==1.0.1
requests==2.32.3
Pillow==10.3.0